refactor: Dialog 코드 네비게이션 되도록 리팩토링
